+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 ## [Unreleased]
 
 - ContentTest - Check length of array or string
-- ContentTest - Check minimum length of array or string
-- ContentTest - Check maximum length of array or string
+- ContentTest - Check minimum length of array or string of response body
+- ContentTest - Check maximum length of array or string of response body
+- ContentTest - Check minimum length of array or string of response header
+- ContentTest - Check maximum length of array or string of response header
 
 ## v1.12.2 - (2021-12-15)